[INJIWEB-1735] Renamed the unknown verifier logo
